# codebase-analysis-docs/assets/api_contract_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Legal Chatbot API",
    "description": "API for querying Indian Constitution and Legal Cases via ChromaDB and Ollama (Qwen3)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root information endpoint",
        "operationId": "root",
        "responses": {
          "200": {
            "description": "API status and docs link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "AI Legal Chatbot API" },
                    "version": { "type": "string", "example": "1.0.0" },
                    "docs": { "type": "string", "example": "/docs" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Check API health and downstream components",
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Health status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server or component check error"
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "summary": "Get database statistics",
        "operationId": "get_stats",
        "responses": {
          "200": {
            "description": "Vector database statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database not initialized or query failure"
          }
        }
      }
    },
    "/api/chat": {
      "options": {
        "summary": "Handle CORS preflight",
        "responses": {
          "200": {
            "description": "Preflight accepted"
          }
        }
      },
      "post": {
        "summary": "Process legal query through RAG pipeline",
        "operationId": "chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated legal answer with context sources",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "500": {
            "description": "Pipeline failure or model communication error"
          }
        }
      }
    },
    "/api/search": {
      "post": {
        "summary": "Search vector database without LLM generation",
        "operationId": "search_documents",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 5 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked search results from ChromaDB",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": { "type": "string" },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchResultItem"
                      }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/build-db": {
      "post": {
        "summary": "Ingest raw dataset into ChromaDB",
        "operationId": "build_database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingestion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "added": { "type": "integer", "example": 1000 }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Source JSON data file not found"
          }
        }
      }
    },
    "/api/doc/{doc_id}": {
      "get": {
        "summary": "Retrieve single document by ID",
        "operationId": "get_document",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Stored document details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "document": { "type": "string", "nullable": true },
                    "metadata": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": { "type": "string", "minLength": 1, "maxLength": 1000 },
          "conversation_id": { "type": "string", "nullable": true },
          "top_k": { "type": "integer", "minimum": 1, "maximum": 10, "default": 5 }
        }
      },
      "ChatResponse": {
        "type": "object",
        "required": ["response", "sources", "conversation_id", "timestamp"],
        "properties": {
          "response": { "type": "string" },
          "sources": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SourceItem" }
          },
          "conversation_id": { "type": "string" },
          "timestamp": { "type": "string" }
        }
      },
      "SourceItem": {
        "type": "object",
        "required": ["title", "source", "category", "url", "preview"],
        "properties": {
          "title": { "type": "string" },
          "source": { "type": "string" },
          "category": { "type": "string" },
          "url": { "type": "string" },
          "preview": { "type": "string" }
        }
      },
      "SearchResultItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "nullable": true },
          "instruction": { "type": "string", "nullable": true },
          "response": { "type": "string", "nullable": true },
          "content_preview": { "type": "string" },
          "score": { "type": "number", "nullable": true }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": ["status", "vector_database_status", "vector_database_count", "ollama_status", "timestamp"],
        "properties": {
          "status": { "type": "string" },
          "vector_database_status": { "type": "string" },
          "vector_database_count": { "type": "integer" },
          "ollama_status": { "type": "string" },
          "timestamp": { "type": "string" }
        }
      },
      "StatsResponse": {
        "type": "object",
        "required": ["total_documents", "categories", "document_types", "collection_name"],
        "properties": {
          "total_documents": { "type": "integer" },
          "categories": { "type": "array", "items": { "type": "string" } },
          "document_types": { "type": "array", "items": { "type": "string" } },
          "collection_name": { "type": "string" }
        }
      },
      "BuildRequest": {
        "type": "object",
        "properties": {
          "reset": { "type": "boolean", "default": false },
          "max_items": { "type": "integer", "nullable": true },
          "filename": { "type": "string", "default": "legal_data_all.json" }
        }
      }
    }
  }
}